requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
comtypes==1.1.7
cryptography==3.2.1
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.4
django-chartit==0.2.9
django-filter==2.3.0
django-import-export==2.1.0
django-mathfilters==1.0.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
fire==0.3.1
idna==2.10
jdcal==1.4.1
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
kombu==4.6.8
lml==0.0.9
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.2.1
mpld3==0.3
nltk==3.5
numpy==1.18.4
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.0.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyexcel==0.6.1
pyexcel-io==0.5.20
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.8
PyMySQL==0.10.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
scipy==1.5.0
seaborn==0.10.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
termcolor==1.1.0
texttable==1.6.2
tqdm==4.48.0
vincent==0.4.4
urllib3==1.25.9
vine==1.3.0
xlrd==1.2.0
XlsxWriter==1.2.8
xlwt==1.3.0
